requirements.txt
streamlit>=1.28.0
anthropic>=0.18.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
plotly>=5.18.0
python-dotenv>=1.0.0
plotly==6.5.2
python-dotenv
